Add set_jobs function
Add a function to set litani's jobs given a list of job dicts. Replace
that functionality in transform_jobs. Move transform_jobs into
lib/jobs.py
